[IMP] base, mail: better support multi emails when computing email_normalized
Previously to this commit setting a multi email on a partner or a lead leads
to a void email_normalized. Indeed default behavior of ``email_normalize``
tool is to enforce the singleton behavior of input text.

This has several drawbacks

  * searching partners or fetching information based on emails does not work as
    most tool methods use ``email_normalized`` which is Falsy (see notably
    ``_message_partner_info_from_emails`` or ``_mail_find_partner_from_emails``)
  * blacklist is completely not available for partners or leads being encoded
    with multi emails (as it is based on ``email_normalized`` which is Falsy);

However we think it is better in some cases to be defensive and support a bit
various inputs that may exist. When computing ``email_normalized`` field it now
takes the first found normalize version of email instead of simply setting
it as Falsy. This has some limitations, notably emails after the first found
one are not taken into account. However this already improves the support while
being a small change and compatible with current code and behavior.

Task-2612945 (Mail: Defensive email formatting)