[IMP] various: improve multi-emails support in post-message hook updating partner
Purpose of this commit is to improve support of multi emails in the post
message hook. In most customer oriented models (Lead, Task, Event, Track,
Ticket) an update is done after posting to link records to the newly created
partner (if any). This is generally used when creating partners on the fly
with suggested recipients in Chatter UX.

In this commit, we now check on either the complete email, either on the
normalized version of email. This allows to get rid of formatting issues
as well as improving multi email support, as comparison will be done on
first found email.

Task-2612945 (Mail: Defensive email formatting)