[IMP] website: add options for file uploading (max size + max number)
Adding options to define the maximum number of files and the maximum
file size (in MB) allowed to be uploaded in a form.
The default values are 1 file and 1 MB.
If the user uploads too many/too big files, the form is not sent and a
message is displayed, informing why it's wrong.
When a user wants to upload files, he needs to know in advance what he
is allowed to do. Therefore, a tooltip message describing the number of
files and the files size limit he can upload is displayed when hovering
on the field.

task-2798576