[IMP] base: replace LXML HTML cleaner by bleach
Purpose
=======
LXML has officially declared that the Cleaner is not appropriate for
untrusted inputs. Because of that we want to refactor our HTML
sanitizer.

Changes
=======
Now, empty tag are automatically closed `<br>` => `<br />`

When sanitize_attributes is set to False, we use an extended whitelist
of safe attributes.

The single quote and double quote are escaped when they are in the
HTML element content. This is wanted as it can be used to trick the
browser parser.

Bleach 4.0.1 sort alphabetically the attributes. This is not the case
anymore with Bleach 5.0 but there's nothing we can do about that in
this bleach version.

The email quote (when you reply to an email, the parent email is
automatically quoted) has been removed from the HTML sanitizer and is
only done on incoming emails. This is useless in most cases for the HTML
sanitation.

Technical
=========
The Bleach sanitizer have an option to strip the not allowed elements.
But it will just remove the element itself and not its children. Also,
we want to have more control on what we do. E.G. we want to allow the
"data:" protocol only for non-svg image and bleach just offer you to
block the entire protocol or not.

For those reason, we need to use directly `BleachSanitizerFilter` and to
overwrite some method to have the behavior we want.

We have a list of "kill tags". When those tags are removed, we also
remove all of their children.

Task-2812488