[FIX] mail: broken email scheduled date
Step to reproduce:
	activate the developer mode
	go to settings - technical - emails
	create a new email and set the Scheduled Send Date in the future
	run the scheduled action Mail: Email Queue Manager

Current behavior:
the email is sent and the action does not consider the filter

Expected behavior:
the email is not sent and will only be sent when the scheduler detects
that scheduled_date exceeds the current time.

Issue description:
The scheduled_date field of the mail is of type char and contains
the time in local time. We need to set "now" with the right format and
timezone for the filter to compare the two correctly.

opw-2823106