[FIX] website: use rerouted url to generate canonical url
Since commit [1], when a website homepage is not the page with the URL
`/`, a traceback would be shown for the public users.

The initial requested path (`/`) is stored inside the `rerouting`
attribute of the request.
After the `reroute()`, the `request.httprequest.path` is `/another_url`,
which shouldn't be used when checking the canonical URL, but instead use
the `request.rerouting` URL `/`.

Indeed, if you chose `/another_url` as homepage (which will be served
when you request `/` through the `@route('/')` controller), it would
ends up raising a `werkzeug.exceptions.NotFound` as that `/another_url`
would not be found in the werkzeug routing map.

Step to reproduce:
- Create a new page with an URL != than `/` (or use the contactus page)
- On that new page (or contactus page), open the page properties dialog
  from the `Pages` menu.
- Go to the second tab in that dialog and toggle the "Use as Homepage"
  option.
- Now, open the `/` URL in public user, a traceback will be shown.

Note that [1] got fined tuned by [2] in the same PR, but those changes
are not relevant in what this commit fixes.

Courtesy of @madprog for the detailed analysis which helped a lot
figuring the issue and coming up with this fix.

opw-2834546
opw-2834549
opw-2834625

[1]: https://github.com/odoo/odoo/commit/80a04f7ebede0d12c185d3190b87eb36f279e141
[2]: https://github.com/odoo/odoo/commit/33167b3928c767a08fdc7eedc3e6204aac9cac08