[IMP] web_editor: add filter-in to m2o/m2m snippet widget
Adds basic support to filter possible m2x values based on another record.

This is necessary for the website_appointment snippet which is introduced
within this bundle (see odoo/enterprise#23750).

The important need implemented here is to not store a list of valid IDs in the
DOM but fetch them based on data stored by a different widget.

Task-2574175
See odoo/enterprise#23750