[FIX] web_editor: give higher priority to image resizing
Before this commit if an image with a `w-100` was resized, its newly
defined style was no taken into account.

After this commit if the style would not be taken into account, then
it is applied as `!important` in order to have more priority than the
CSS class.

This commit must not be forwarded beyond 14.0.

task-2841127