[IMP] core: utility to reraise errors as others
We introduce a new decorator/context-manager to catch some exceptions
and re-raise them as fake alternative error. This utility main's purpose
is to hide a route that a user has no access to behind a fake HTTP 404
Page not Found error.

The utility is at `odoo.tools.misc.reraise_x_as(*exceptions, as_)`.

Its usage is as follow:

    @route('/some/route', auth='public')
    @reraise_x_as(AccessError, AccessDenied, as_=NotFound())
    def some_route(self):
        if not request.session.uid:
            raise AccessError("Must be connected to see this route")
        ...

Or as a context-manager if you don't want to except an entire function:

    @route('/some/route', auth='public')
    def some_route(self):
        with reraise_x_as(AccessError, AccessDenied, as_=NotFound()):
            if not request.session.uid:
                raise AccessError("Must be connected to see this route")
        ...

Task: 2800772
Close: #90433