[IMP] survey: allow to share survey results with anyone
PURPOSE:
The results page of a survey is now "anyone with the link"-based access.

When clicking the "See Results" action on the survey, the URL can now
be shared and used by anyone, even by public users.

A randomly generated "results_token" has been added to the survey model,
so that the result page is still kind of hidden (the URL can not be rebuilt
using other info such as the survey's name or the survey's access token).

Task-2842682