[FIX] website: change some snippets default background color
Some snippets have a "hard coded" background color which doesn't change
when we want to put a preset color (theme). We have to delete the back-
ground first before being able to choose a preset color.

This commit fixes this behaviour by putting a preset color by default
on the concerned snippets (instead of a fixed color).

The snippets mentionned in the task are searchbar, cover and parallax.
But an other one (text highlight) was found to have the same behaviour
and was also fixed.

task-2824393