[FIX] sale_stock: Cancel a SO with non delivery picking done
In the case of a multi-step delivery, a warning would be thrown telling
goods have already been delivered when only one step (like packing) was
done, which is wrong.

In the case of a MTO MO flow, just validating the 'pick components'
picking for the MO would trigger this warning as well.

Task-2797359