[FIX] account, test_sale_product_configurators: fixing currency issues
test_sale_product_configurators relied on the fact that demo
data is installed with multi_currency, but that's only because
in the old demo data, the company is created with EUR, which is made
active by the sale module, and then changed to USD, and that
turns on the multicurrency group, that activates the
group_product_pricelist. But now, as USD is set before sale is
installed, only USD is active, and the group_product_pricelist is not
default. So we have to add it in the test.