[IMP] *: provide a parameter to disable image quality optimization
This commit allows to have a config parameter (not set by default) that
allows to disable the quality optimization of the images that are
uploaded by the user. From this commit, if there is a value for the
Ir.config.parameter 'global.no_quality_optimization', the optimization
of images will be disabled. This remains a hidden function in the system
parameters.

task-2835144