[IMP] *: provide a param to disable website image quality optimization
This commit allows to have a config parameter (not set by default) that
allows to disable the image quality optimization when the user edits his
website. From this commit, if there is a value for the
Ir.config.parameter 'website.no_quality_optimization', the image
optimization will be disabled on the website builder. This remains a
hidden function in the system parameters.

task-2835144