[FIX] mass_mailing_sms: make retry generate cron triggers
When we use the "retry" button in case of a failed mailing, it puts the mailing
back to the queue where it will be sent on the next cron trigger (generally on
the next day by the  `Mail Marketing: Process queue` cron).

After this commit, the mailing will be scheduled to be sent at the next
`SMS: SMS Queue Manager` cron trigger.

taskID-2760101