[FIX] point_of_sale: avoid negative discounts on tickets and invoices
The discounts printed on point of sale tickets and invoices do not correctly
take the discount policy into account for recursive pricelists. This commit
adapts the discounts shown on invoices generated by the point of sale to mimic
those on sale order invoices.

However, on the js side, currently no code is available to calculate the
correct list price when a base pricelist with discounts included is involved.
To prevent the most flagrant anomalies, at least a check is introduced to
filter out discounts that make no sense when calculating the total discount.

opw-2834291